// frontend/src/types/bowler.ts
// shadows backend model for food "MarriottFood"
//lowercase
export type Bowler = {
    bowlerID: number;
    bowlerLastName: string;
    bowlerFirstName: string;
    bowlerMiddleInit?: string;
    bowlerAddress: string;
    teamID: number;
    bowlerCity: string;
    bowlerState: string;
    bowlerZip: string;
    bowlerPhoneNumber: string;
  };
  